# backendnest/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestão Financeira Pessoal
  description: |
    API REST para autenticação e gestão de contas, categorias, transações,
    transferências, dívidas, metas e alertas.

    ## Convenção de idioma
    Os nomes dos campos seguem **camelCase em português**, pois as entidades
    do banco de dados estão em português (ex. `nome`, `senha`, `saldoInicial`).
    Alguns campos de domínio financeiro estão em espanhol (ex. `montoTotal`,
    `fechaVencimiento`, `diasAnticipacion`).

    ## Autenticação
    Todos os endpoints (exceto `/auth/register` e `/auth/login`) requerem um
    token JWT no header `Authorization: Bearer <token>`.
    O token expira em **1 hora**.

    ## Códigos de status
    - `200` — Operação bem-sucedida (GET, PATCH, DELETE)
    - `201` — Recurso criado (POST)
    - `400` — Validação de campos falhou
    - `401` — Token ausente, expirado ou inválido
    - `404` — Recurso não encontrado ou não pertence ao usuário
    - `409` — Conflito (ex. e-mail duplicado)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido em /auth/login. Expira em 1 hora.
  schemas:
...
    RegisterRequest:
      type: object
      required: [nome, email, senha]
      properties:
        nome:
          type: string
          description: Nome completo do usuário.
          example: Andre
        email:
          type: string
          format: email
          description: E-mail único do usuário (usado para login).
          example: andre@test.com
        senha:
          type: string
          minLength: 6
          description: Senha (mínimo 6 caracteres). Armazenada como hash bcrypt.
          example: "123456"
    RegisterResponse:
      type: object
      properties:
        usuario:
          type: object
          description: Dados públicos do usuário recém-criado.
          properties:
            id:
              type: string
              format: uuid
            nome:
              type: string
            email:
              type: string
    LoginRequest:
      type: object
      required: [email, senha]
      properties:
        email:
          type: string
          format: email
          description: E-mail registrado.
          example: andre@test.com
        senha:
          type: string
          description: Senha em texto plano (comparada com o hash).
          example: "123456"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Token JWT válido por 1 hora. Enviar no header Authorization.
        usuario:
          type: object
          description: Dados públicos do usuário autenticado.
          properties:
            id:
              type: string
              format: uuid
            nome:
              type: string
            email:
              type: string
    ResetPasswordRequest:
      type: object
      required: [novaSenha]
      properties:
        novaSenha:
          type: string
          minLength: 6
          description: Nova senha (mínimo 6 caracteres). O e-mail é obtido do token.
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensagem de confirmação da operação.
          example: Senha atualizada com sucesso

    # ══════════════════════════════════════
    #  CONTAS
    # ══════════════════════════════════════
    CreateContaRequest:
      type: object
      required: [nome, tipo, saldoInicial]
      properties:
        nome:
          type: string
          description: Nome descritivo da conta.
          example: Carteira
        tipo:
          type: string
          enum: [dinheiro, banco, poupanca, cartao_credito]
          description: |
            Tipo de conta:
            - `dinheiro` — Dinheiro em espécie
            - `banco` — Conta bancária
            - `poupanca` — Conta poupança
            - `cartao_credito` — Cartão de crédito
        saldoInicial:
          type: number
          format: double
          description: Saldo inicial da conta (decimal 12,2).
          example: 500.00
        limiteCredito:
          type: number
          format: double
          nullable: true
          description: Limite de crédito (aplica-se apenas para cartão de crédito).
        dataCorte:
          type: integer
          nullable: true
          description: Dia do mês de corte (1-31). Apenas para cartão de crédito.
        dataPagamento:
          type: integer
          nullable: true
          description: Dia do mês de pagamento (1-31). Apenas para cartão de crédito.
    UpdateContaRequest:
      type: object
      description: Apenas os campos enviados serão atualizados. Não é permitido alterar `tipo` nem `saldoInicial`.
      properties:
        nome:
          type: string
          description: Novo nome da conta.
        limiteCredito:
          type: number
          format: double
          description: Novo limite de crédito.
        dataCorte:
          type: integer
          description: Novo dia de corte.
        dataPagamento:
          type: integer
          description: Novo dia de pagamento.
        ativa:
          type: boolean
          description: Estado ativo/inativo da conta.
    Conta:
      type: object
      description: Entidade completa de uma conta financeira.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único da conta.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono da conta.
        nome:
          type: string
          description: Nome descritivo.
        tipo:
          type: string
          enum: [dinheiro, banco, poupanca, cartao_credito]
          description: Tipo de conta financeira.
        saldoInicial:
          type: number
          format: double
          description: Saldo com o qual a conta foi criada.
        moeda:
          type: string
          description: Moeda da conta (padrão BRL).
          example: BRL
        limiteCredito:
          type: number
          nullable: true
          description: Limite de crédito (null se não se aplica).
        dataCorte:
          type: integer
          nullable: true
          description: Dia de fechamento da fatura (null se não se aplica).
        dataPagamento:
          type: integer
          nullable: true
          description: Dia de vencimento da fatura (null se não se aplica).
        ativa:
          type: boolean
          description: Se a conta está ativa (false = soft delete).
        createdAt:
          type: string
          format: date-time
          description: Data de criação.
        updatedAt:
          type: string
          format: date-time
          description: Data da última atualização.

    # ══════════════════════════════════════
    #  CATEGORIAS
    # ══════════════════════════════════════
    CreateCategoriaRequest:
      type: object
      required: [nome, tipo]
      properties:
        nome:
          type: string
          description: Nome da categoria.
          example: Alimentação
        tipo:
          type: string
          enum: [receita, despesa]
          description: |
            Tipo de categoria:
            - `receita` — Receita (entrada)
            - `despesa` — Despesa (saída)
        cor:
          type: string
          nullable: true
          description: Cor hexadecimal para representação visual.
          example: "#FF5733"
        icone:
          type: string
          nullable: true
          description: Nome do ícone para a interface.
          example: utensils
    UpdateCategoriaRequest:
      type: object
      description: Apenas os campos enviados serão atualizados. Não é permitido alterar `tipo`.
      properties:
        nome:
          type: string
          description: Novo nome.
        cor:
          type: string
          description: Nova cor hexadecimal.
        icone:
          type: string
          description: Novo ícone.
        ativa:
          type: boolean
          description: Estado ativo/inativo.
    Categoria:
      type: object
      description: Entidade completa de uma categoria.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        nome:
          type: string
          description: Nome da categoria.
        tipo:
          type: string
          enum: [receita, despesa]
          description: Tipo (receita ou despesa).
        cor:
          type: string
          nullable: true
          description: Cor hexadecimal.
        icone:
          type: string
          nullable: true
          description: Nome do ícone.
        ativa:
          type: boolean
          description: Se está ativa.
        createdAt:
          type: string
          format: date-time
          description: Data de criação.

    # ══════════════════════════════════════
    #  TRANSACOES
    # ══════════════════════════════════════
...
    CreateAlertaRequest:
      type: object
      required: [tipo, referenciaId, diasAnticipacion]
      properties:
        tipo:
          type: string
          enum: [vencimento_divida, vencimento_meta, limite_gasto]
          description: |
            Tipo de alerta:
            - `vencimento_divida` — Aviso de vencimento de dívida
            - `vencimento_meta` — Aviso de vencimento de meta
            - `limite_gasto` — Aviso de limite de gasto atingido
        referenciaId:
          type: string
          format: uuid
          description: ID do recurso ao qual o alerta se refere (dívida, meta, etc.).
        diasAnticipacion:
          type: integer
          minimum: 1
          description: Dias de antecedência para disparar o alerta.
          example: 7
    UpdateAlertaRequest:
      type: object
      properties:
        diasAnticipacion:
          type: integer
          minimum: 1
          description: Novos dias de antecedência.
        ativa:
          type: boolean
          description: Estado ativo/inativo.
    Alerta:
      type: object
      description: Entidade completa de um alerta.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        tipo:
          type: string
          enum: [vencimento_divida, vencimento_meta, limite_gasto]
          description: Tipo de alerta.
        referenciaId:
          type: string
          format: uuid
          description: ID do recurso referenciado.
        diasAnticipacion:
          type: integer
          description: Dias de antecedência configurados.
        ativa:
          type: boolean
          description: Se o alerta está ativo.
        ultimaNotificacion:
          type: string
          format: date-time
          nullable: true
          description: Data/hora da última notificação enviada (null se nunca foi notificado).
        createdAt:
          type: string
          format: date-time
          description: Data de criação.

    # ══════════════════════════════════════
    #  ERROS
    # ══════════════════════════════════════
    Error400:
      type: object
      description: Erro de validação. O campo `message` pode ser string ou array de strings com os erros específicos.
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Detalhe do erro de validação.
        error:
          type: string
          example: Bad Request
    Error401:
      type: object
      description: Erro de autenticação (token ausente, expirado ou inválido).
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
    Error404:
      type: object
      description: Recurso não encontrado ou não pertence ao usuário autenticado.
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Conta não encontrada
    Error409:
      type: object
      description: Conflito (ex. e-mail já cadastrado).
      properties:
        statusCode:
          type: integer
          example: 409
        message:
          type: string
          example: E-mail já cadastrado
  responses:
    BadRequest:
      description: Validação de campos falhou.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error400'
    Unauthorized:
      description: Token JWT ausente, expirado ou inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error401'
    NotFound:
      description: Recurso não encontrado ou não pertence ao usuário autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error404'
    Conflict:
      description: Conflito (recurso duplicado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error409'
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar novo usuário
      description: Cria um novo usuário. O e-mail deve ser único. A senha é armazenada como hash bcrypt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Fazer login
      description: |
        Autentica o usuário e retorna um token JWT válido por 1 hora.
        O token deve ser enviado no header `Authorization: Bearer <token>` em todas as demais requisições.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login bem-sucedido. Retorna token JWT e dados do usuário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: E-mail ou senha inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Alterar senha
      description: |
        Altera a senha do usuário autenticado. O e-mail é obtido automaticamente do token JWT,
        portanto só é necessário enviar a nova senha no body.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '201':
          description: Senha atualizada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════
  #  CONTAS
  # ════════════════════════════════════════
  /contas:
    post:
      tags: [Contas]
      summary: Criar conta
      description: Cria uma nova conta financeira para o usuário autenticado.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContaRequest'
      responses:
        '201':
          description: Conta criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Contas]
      summary: Listar contas ativas
      description: Retorna todas as contas **ativas** (`ativa = true`) do usuário autenticado.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de contas ativas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /contas/{id}:
    get:
      tags: [Contas]
      summary: Obter conta por ID
      description: Retorna uma conta específica do usuário autenticado.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da conta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe da conta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Contas]
      summary: Atualizar conta
      description: Atualiza os campos enviados de uma conta. Retorna a conta atualizada.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da conta.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContaRequest'
      responses:
        '200':
          description: Conta atualizada. Retorna a entidade completa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contas/{id}/desativar:
    patch:
      tags: [Contas]
      summary: Desativar conta (soft delete)
      description: Marca a conta como inativa (`ativa = false`). Não exclui dados. Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da conta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conta desativada com sucesso. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  CATEGORIAS
  # ════════════════════════════════════════
  /categorias:
    post:
      tags: [Categorias]
      summary: Criar categoria
      description: Cria uma nova categoria de receita ou despesa.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoriaRequest'
      responses:
        '201':
          description: Categoria criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Categorias]
      summary: Listar categorias ativas
      description: Retorna todas as categorias **ativas** do usuário autenticado.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de categorias ativas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /categorias/{id}:
    get:
      tags: [Categorias]
      summary: Obter categoria por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da categoria.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe da categoria.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Categorias]
      summary: Atualizar categoria
      description: Atualiza os campos enviados. Retorna a categoria atualizada.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da categoria.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoriaRequest'
      responses:
        '200':
          description: Categoria atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /categorias/{id}/desativar:
    patch:
      tags: [Categorias]
      summary: Desativar categoria (soft delete)
      description: Marca como inativa. Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da categoria.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Categoria desativada. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  TRANSACOES
  # ════════════════════════════════════════
...
  /alertas:
    post:
      tags: [Alertas]
      summary: Criar alerta
      description: Cria um novo alerta vinculado a uma dívida, meta ou limite de gasto.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertaRequest'
      responses:
        '201':
          description: Alerta criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alerta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Alertas]
      summary: Listar alertas ativos
      description: |
        Retorna todos os alertas **ativos** (`ativa = true`) do usuário autenticado,
        ordenados por data de criação **descendente**.
        Atualmente não suporta filtros por query params.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de alertas ativos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alerta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /alertas/{id}:
    get:
      tags: [Alertas]
      summary: Obter alerta por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do alerta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe do alerta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alerta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Alertas]
      summary: Atualizar alerta
      description: Atualiza os campos enviados. Retorna o alerta atualizado.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do alerta.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertaRequest'
      responses:
        '200':
          description: Alerta atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alerta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /alertas/{id}/desativar:
    patch:
      tags: [Alertas]
      summary: Desativar alerta
      description: Marca o alerta como inativo. Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do alerta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alerta desativado. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /alertas/{id}/notificar:
    patch:
      tags: [Alertas]
      summary: Marcar alerta como notificado
      description: |
        Registra a data/hora atual em `ultimaNotificacion`.
        Útil para que o frontend saiba quando foi notificado pela última vez. Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do alerta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alerta marcado como notificado. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'